Rename DirectSpeakerGainCalc files to DirectSpeakersGainCalc
This aligns it with the class name and the comments in the files.
